SD Format

+++ SilverDasher/SilverDasher/ACT/Models/Coordinate.cs
@@ -2,17 +2,12 @@ using Newtonsoft.Json;
 
 namespace SilverDasher.ACT.Models;
 
-internal class Coordinate
-{
-	[JsonProperty("x")]
-	internal int x;
+internal class Coordinate {
+    [JsonProperty("x")] internal int x;
 
-	[JsonProperty("y")]
-	internal int y;
+    [JsonProperty("y")] internal int y;
 
-	[JsonIgnore]
-	internal string displayX => ((float)x / 100f).ToString("0.##");
+    [JsonIgnore] internal string displayX => (x / 100f).ToString("0.##");
 
-	[JsonIgnore]
-	internal string displayY => ((float)y / 100f).ToString("0.##");
+    [JsonIgnore] internal string displayY => (y / 100f).ToString("0.##");
 }
